core/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.25
django-filter==23.5
djangorestframework==3.15.1
idna==3.7
importlib_metadata==8.4.0
itypes==1.2.0
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==2.1.5
pillow==10.4.0
python-decouple==3.8
pytz==2024.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
zipp==3.20.0
